tried to use conversion to double for both int and double (optional) - the if (data) cast to bool does not work, I'm drowning and drowning

+++ CeTu.h
@@ -73,9 +73,15 @@ namespace CeTu
     {
     public:
 
-        coutt& operator << (int val)
+//        coutt& operator << (int val)
+//        {
+//            printf("%d", val);
+//            return (*this);
+//        }
+
+        coutt& operator << (double val)
         {
-            printf("%d", val);
+            printf("%f", val);
             return (*this);
         }
